+++ src/sched/entry/coll/direct/bcast_entry.hpp
@@ -28,7 +28,7 @@ public:
                 ccl_buffer buf,
                 size_t cnt,
                 const ccl_datatype& dtype,
-                size_t root,
+                int root,
                 ccl_comm* comm)
             : base_coll_entry(sched),
               buf(buf),
@@ -92,7 +92,7 @@ protected:
 private:
     ccl_buffer buf;
     size_t cnt;
-    size_t root;
+    int root;
     ccl_datatype dtype;
     ccl_comm* comm;
     atl_req_t req{};
